fix: destroy time as float values

+++ src/models.py
@@ -7,7 +7,7 @@ class RequestLink(BaseModel):
     keyword: Optional[str] = ''
     tags: Optional[list] = []
     destroy_clicks: Optional[int] = 0
-    destroy_time: Optional[int] = 0
+    destroy_time: Optional[float] = 0
 
 
 class Keyword(BaseModel):
@@ -21,4 +21,4 @@ class ResponseLink(BaseModel):
     destroy_clicks: int
     tags: list
     date_created: float
-    destroy_time: int
+    destroy_time: float
